Allow user override of GOPATH
Repeat building for developers can benefit from being able to shortcut
full executions through cache reuse. Allow user to override the GOPATH
used by specifying a module specific environment variable. This can
allow a user to reuse a pre-populated golang module and build cache for
faster compiles when developing.

Note it requires a valid gomodule package, therefore re-using the
gomodules test to validate that the GOPATH is modified.

Local testing indicates that for a module in a custom project the
install time could be reduced from 35 seconds to 9 seconds. While not
perfect, it is still useful for local development environments,
especially if more modules are used.